% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLenDataBase-pkg.R
\docType{package}
\name{geneLenDatabase-pkg}
\alias{geneLenDataBase}
\alias{geneLenDataBase-package}
\alias{geneLenDatabase-pkg}
\title{geneLenDatabase:}
\description{
Lengths of mRNA transcripts for a number of genomes
}
\details{
Length of mRNA transcripts for a number of genomes and gene ID formats, 
largely based on UCSC table browser.
Data objects are provided as individual pieces of information to be retrieved 
and loaded. 
A variety of different gene identifiers and genomes is supported to ensure 
wide applicability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/federicomarini/geneLenDataBase}
  \item Report bugs at \url{https://github.com/federicomarini/geneLenDataBase/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID}) [contributor]

Authors:
\itemize{
  \item Matthew Young \email{my4@sanger.ac.uk}
  \item Nadia Davidson \email{nadia.davidson@mcri.edu.au}
}

}
\keyword{internal}
